Do some recommended refactoring
* Make a single helper for parsing JSON
* Remake a helper for getting user-to-conversation mappings to save on
  repetitive code
